Break up fuellib/__init__
