+++ docs/sourcecode.rst
@@ -16,28 +16,31 @@ FuelLib File Organization
 - **gcmTableData:** directory that contains the pre-tabulated group contributions
 - **fuellib:** main package directory containing:
 
-    - ``FuelLib.py``: class for enabling GCM predictions
+    - ``fuel.py``: core :class:`fuel` class for Group Contribution Method calculations
+    - ``constants.py``: physical constants (Boltzmann, Avogadro)
+    - ``conversions.py``: temperature conversion functions and Lennard-Jones calculations
+    - ``utilities.py``: utility functions for mixture properties and droplet calculations
+    - ``_data_locator.py``: internal module for locating and validating fuel data directories
     - **exporters:** subpackage with CLI exporters for generating fuel properties
     
         - ``converge.py``: exporter for Converge CFD simulations (CLI: ``fl-export-converge``)
         - ``pele.py``: exporter for PelePhysics simulations (CLI: ``fl-export-pele``)
 
-    - **utilities:** subpackage with CLI utility tools for data conversion and analysis
+    - **cli:** subpackage with command-line interface tools for data conversion and analysis
     
-        - ``transport_props_converter.py``: convert Lennard-Jones well depth from J/mol to characteristic temperature in Kelvin (CLI: ``fl-eps2K``)
-        - ``temp_converter.py``: convert temperature between Celsius and Kelvin (CLI: ``fl-C2K``, ``fl-K2C``)
-        - ``plotting.py``: plot fuel composition and mixture properties
-        
-            - ``plot_composition()``: plot fuel composition by compound and chemical family (CLI: ``fl-plt-comp``)
-            - ``plot_mixture_properties()``: plot mixture properties over temperature range with optional experimental data (CLI: ``fl-plt-props``)
+        - ``temp_converter.py``: temperature conversion utilities (CLI: ``fl-C2K``, ``fl-K2C``, ``fl-C2F``, ``fl-F2C``, ``fl-F2K``, ``fl-K2F``)
+        - ``transport_props_converter.py``: Lennard-Jones parameter conversion (CLI: ``fl-eps2K``)
+        - ``plotting.py``: plotting utilities for composition and properties (CLI: ``fl-plt-comp``, ``fl-plt-props``)
+        - ``_fuel_manager.py``: fuel manager utility (CLI: ``fl-fuels``)
 
 - **tests:**  directory containing CI unit tests for FuelLib. The CI test checks if the cumulative error of property predictions of a new proposed model are less than or equal to the current model.
     
     - **baselinePredictions:** directory that contains baseline predictions and script ``generate_baseline.py`` for generating baseline predictions for CI testing.
     - ``test_accuracy.py``: unit test used in CI for verifying new model predictions preserve accuracy
     - ``test_source_docstrings.py``: documentation contract test that checks public source functions include required docstring fields (``:param:``, ``:type:``, ``:return:``, ``:rtype:``).
-    - ``test_api.py``: combined API/signature and function-evaluation test that checks public ``FuelLib.py`` module and class method signatures for unexpected API drift and runs representative FuelLib smoke evaluations.
+    - ``test_api.py``: combined API/signature and function-evaluation test that checks public fuellib module and class method signatures for unexpected API drift and runs representative FuelLib smoke evaluations.
     - ``test_utilities.py``: unit test for utility functions and CLI commands including temperature conversion and transport property calculations.
+    - ``test_hc_identification.py``: unit test for hydrocarbon classification logic.
     - ``get_pred_and_data.py``: helper function used by ``test_accuracy.py`` and ``baselinePredictions/generate_baseline.py`` to compute predictions and load validation data.
 
 - **tutorials:** directory containing example scripts that demonstrate how to use FuelLib
