+++ fuellib/cli/_fuel_manager.py
