+++ fuellib/cli/plotting.py
