+++ fuellib/constants.py
@@ -1,0 +1,7 @@
+"""Physical constants used in FuelLib calculations."""
+
+# Physical constants
+k_B = 1.380649e-23  # Boltzmann's constant in J/K
+N_A = 6.02214076e23  # Avogadro's number in 1/mol
+
+__all__ = ["k_B", "N_A"]
