+++ pyproject.toml
@@ -51,16 +51,16 @@ fl-export-pele = "fuellib.exporters.pele:main"
 fl-build-docs = "fuellib.build_docs:main"
 fl-clean-docs = "fuellib.clean_docs:main"
 fl-format = "fuellib.format_code:main"
-fl-eps2K = "fuellib.utilities.transport_props_converter:eps2K_main"
-fl-C2K = "fuellib.utilities.temp_converter:c2k_main"
-fl-K2C = "fuellib.utilities.temp_converter:k2c_main"
-fl-C2F = "fuellib.utilities.temp_converter:c2f_main"
-fl-F2C = "fuellib.utilities.temp_converter:f2c_main"
-fl-F2K = "fuellib.utilities.temp_converter:f2k_main"
-fl-K2F = "fuellib.utilities.temp_converter:k2f_main"
-fl-plt-comp = "fuellib.utilities.plotting:comp_main"
-fl-plt-props = "fuellib.utilities.plotting:props_main"
-fl-fuels = "fuellib.utilities.fuel_manager:list_fuels_main"
+fl-eps2K = "fuellib.cli.transport_props_converter:eps2K_main"
+fl-C2K = "fuellib.cli.temp_converter:c2k_main"
+fl-K2C = "fuellib.cli.temp_converter:k2c_main"
+fl-C2F = "fuellib.cli.temp_converter:c2f_main"
+fl-F2C = "fuellib.cli.temp_converter:f2c_main"
+fl-F2K = "fuellib.cli.temp_converter:f2k_main"
+fl-K2F = "fuellib.cli.temp_converter:k2f_main"
+fl-plt-comp = "fuellib.cli.plotting:comp_main"
+fl-plt-props = "fuellib.cli.plotting:props_main"
+fl-fuels = "fuellib.cli._fuel_manager:list_fuels_main"
 
 [project.urls]
 Repository = "https://github.com/NatLabRockies/FuelLib"
@@ -68,7 +68,7 @@ Issues = "https://github.com/NatLabRockies/FuelLib/issues"
 Documentation = "https://NatLabRockies.github.io/FuelLib"
 
 [tool.setuptools]
-packages = ["fuellib", "fuellib.exporters", "fuellib.utilities"]
+packages = ["fuellib", "fuellib.exporters", "fuellib.cli"]
 
 [tool.setuptools.package-data]
 fuellib = [
